% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_higher_taxon_bracket.R
\name{check_higher_taxon_bracket}
\alias{check_higher_taxon_bracket}
\title{Check if there is only one taxonomic group below specified level}
\usage{
check_higher_taxon_bracket(taxa)
}
\arguments{
\item{taxa}{A character vector of primate taxa.}
}
\value{

}
\description{
Check if there is only one taxonomic group below specified level
}
\examples{
check_higher_taxon_bracket("lepilemur")
check_higher_taxon_bracket("otolemur")
}
