% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_psicinfo_index.R
\name{format_psicinfo_index}
\alias{format_psicinfo_index}
\title{Format PscicInfo index search terms}
\usage{
format_psicinfo_index(taxa)
}
\arguments{
\item{taxa}{A character vector of primate taxa.}
}
\value{
A string of index search terms that are associated with the specified taxa, formatted for use in PscicInfo.
}
\description{
Format PscicInfo index search terms
}
\examples{
format_psicinfo_index("cercopithecidae")
}
