% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_psicinfo_tiab.R
\name{format_psicinfo_tiab}
\alias{format_psicinfo_tiab}
\title{Format PscicInfo title and abstract search terms}
\usage{
format_psicinfo_tiab(taxa)
}
\arguments{
\item{taxa}{A character vector of primate taxa.}
}
\value{
A string of search terms that are associated with the specified taxa, formatted for use in PscicInfo.
}
\description{
Format PscicInfo title and abstract search terms
}
\examples{
format_psicinfo_tiab("lorisidae")
}
