% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_psycinfo_terms.R
\name{format_psycinfo_terms}
\alias{format_psycinfo_terms}
\title{Format PscicInfo index search terms}
\usage{
format_psycinfo_terms(taxa, exclude = NULL)
}
\arguments{
\item{taxa}{A character vector of primate taxa.}

\item{exclude}{A character vector of primate taxa that are higher in the taxonomic tree than those provided in the taxa argument to exclude from the search terms.}
}
\value{
A string of search terms that are associated with the specified taxa, formatted for use in PscicInfo.
}
\description{
Format PscicInfo index search terms
}
\examples{
format_psycinfo_terms("cercopithecidae")
format_psycinfo_terms("cercopithecidae", exclude = "papio")
}
