% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pubmed_mesh.R
\name{format_pubmed_mesh}
\alias{format_pubmed_mesh}
\title{Format PubMed mesh terms}
\usage{
format_pubmed_mesh(taxa)
}
\arguments{
\item{taxa}{A character vector of primate taxa.}
}
\value{
A character vector of mesh search terms that are associated with the specified taxa, formatted for use in PubMed.
}
\description{
Format PubMed mesh terms
}
\examples{
format_pubmed_mesh("papio")
format_pubmed_mesh(c("papio", "macaca"))
}
