% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pubmed_terms.R
\name{format_pubmed_terms}
\alias{format_pubmed_terms}
\title{Format PubMed search terms}
\usage{
format_pubmed_terms(taxa, exclude = NULL)
}
\arguments{
\item{taxa}{A character vector of primate taxa.}

\item{exclude}{A character vector of primate taxa that are higher in the taxonomic tree than those provided in the taxa argument to exclude from the search terms.}
}
\value{
A string of search terms that are associated with the specified taxa, formatted for use in PubMed.
}
\description{
Format PubMed search terms
}
\examples{
format_pubmed_terms("papio")
format_pubmed_terms("cercopithecidae", exclude = "papio")
}
