% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pubmed_tiab.R
\name{format_pubmed_tiab}
\alias{format_pubmed_tiab}
\title{Format PubMed title and abstract terms}
\usage{
format_pubmed_tiab(taxa)
}
\arguments{
\item{taxa}{A character vector of primate taxa.}
}
\value{
A character vector of search terms for the title and abstract that are associated with the specified taxa, formatted for use in PubMed.
}
\description{
Format PubMed title and abstract terms
}
\examples{

format_pubmed_tiab("papio")
format_pubmed_tiab(c("papio", "macaca"))
}
