% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_wos_general_terms.R
\name{format_wos_general_terms}
\alias{format_wos_general_terms}
\title{Format WebOfScience general search terms}
\usage{
format_wos_general_terms(taxa)
}
\arguments{
\item{taxa}{A character vector of primate taxa.}
}
\value{
A string of search terms that are associated with the specified taxa, formatted for use in WebOfScience.
}
\description{
Format WebOfScience general search terms
}
\examples{
format_wos_general_terms("papio")
}
