% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_wos_terms.R
\name{format_wos_terms}
\alias{format_wos_terms}
\title{Format WebOfScience search terms}
\usage{
format_wos_terms(taxa, exclude = NULL)
}
\arguments{
\item{taxa}{A character vector of primate taxa.}

\item{exclude}{A character vector of primate taxa that are higher in the taxonomic tree than those provided in the taxa argument to exclude from the search terms.}
}
\value{
A string of search terms that are associated with the specified taxa, formatted for use in WebOfScience.
}
\description{
Format WebOfScience search terms
}
\examples{
format_wos_terms("cercopithecidae")
format_wos_terms("cercopithecidae", exclude = "papio")
}
