% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_nhp.R
\name{search_nhp}
\alias{search_nhp}
\title{Format non-human primate search terms for use in databases}
\usage{
search_nhp(database = "PubMed", taxa = NULL, exclude = NULL)
}
\arguments{
\item{database}{A string indicating which database search terms should be formated for. Current options are "PubMed" (default), "PsycInfo" or "WebOfScience".}

\item{taxa}{A character vector of primate taxa.}

\item{exclude}{A character vector of primate taxa that are higher in the taxonomic tree than those provided in the taxa argument to exclude from the search terms.}
}
\value{
A string of search terms that are associated with the specified taxa, formatted for use in the specified database.
}
\description{
Format non-human primate search terms for use in databases
}
\examples{
search_nhp(database = "PsycInfo", taxa = "papio")
search_nhp(database = "PubMed", taxa = "cercopithecidae", exclude = "papio")
search_nhp(database = "WebOfScience", taxa = "Platyrrhini", exclude = "Aotus")
}
